requirements.txt
rawpy
mmcv-full==1.3.8
mmgenerataion==0.5.0